+++ start/src/App.vue
@@ -7,6 +7,15 @@
     FIconClock
   } from '@fighting-design/fighting-icon'
 
+  const show1 = ref(false)
+
+  const listImg = [
+    'https://tianyuhao.cn/images/auto/1.jpg',
+    'https://tianyuhao.cn/images/auto/2.jpg',
+    'https://tianyuhao.cn/images/auto/3.jpg',
+    'https://tianyuhao.cn/images/auto/4.jpg'
+  ]
+
   // const value = ref(17)
 
   const num = ref(123)
@@ -25,6 +34,9 @@
 </script>
 
 <template>
+  <f-button type="primary" @click="show1 = true">打开</f-button>
+  <f-image-preview v-model:visible="show1" :img-list="listImg" />
+
   <f-toolbar :on-click="click">
     <f-toolbar-item :index="1" :icon="FIconApps" />
     <f-toolbar-item :index="2" :icon="FIconBlock" />